% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperSpec_conversion.R
\name{as.hyperSpec}
\alias{as.hyperSpec}
\title{hyperSpec}
\usage{
as.hyperSpec(object)
}
\arguments{
\item{object}{Object of class \code{Speclib}.}
}
\value{
Object of class \code{hyperSpec}.
}
\description{
Conversion from \code{Speclib}- to \code{hyperSpec}-object
}
\note{
Package hyperSpec must be installed.
}
\seealso{
\code{\linkS4class{Speclib}}
}
\author{
Lukas Lehnert
}
\keyword{utilities}
