\name{as.hyperSpec}
\alias{as.hyperSpec}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
hyperSpec
}
\description{
Conversion from \code{Speclib}- to \code{hyperSpec}-object. The \code{hyperSpec}-class is the most important class in package \pkg{hyperSpec} which provides methods to analyze spectroscopic datasets.
}
\usage{
as.hyperSpec(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
Object of class \code{Speclib}.
}
}

\value{
Object of class \code{hyperSpec}.
}
\author{
Lukas Lehnert
}
\note{Package \pkg{hyperSpec} must be installed.}

\seealso{
\code{\linkS4class{Speclib}}
}
\keyword{utilities}
