% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speclib_bandnames.R
\name{bandnames}
\alias{bandnames}
\alias{bandnames<-}
\title{Handling names of bands}
\usage{
bandnames(x)
}
\arguments{
\item{x}{Object of class \code{Speclib}.}

\item{value}{Character vector of the same length as \code{nbands(x)}, or
NULL.}
}
\value{
For \code{bandnames<-}, the updated object. Otherwise a vector
giving the name of each band in \code{Speclib} is returned.
}
\description{
Returning and setting names of bands in \code{Speclib}
}
\examples{

data(spectral_data)

bandnames(spectral_data)

}
\seealso{
\code{\linkS4class{Speclib}}
}
\author{
Lukas Lehnert
}
\keyword{utilities}
