% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuum_features.R
\name{bdri}
\alias{bdri}
\title{Band depth ratio indices}
\usage{
bdri(x, fnumber, index = "ndbi")
}
\arguments{
\item{x}{Object of class \code{Specfeat}.}

\item{fnumber}{Integer. Index of feature to modify.}

\item{index}{Method to be applied. Currently, \code{"bdr"}, \code{"ndbi"}
and \code{"bna"} are available.}
}
\value{
Object of class \code{specfeat} containing the updated version of x.
}
\description{
Calculate band depth ratio indices for objects of class \code{Specfeat}.
}
\details{
Method \code{"bdr"} calculates the normalised band depth ratio as
\deqn{bdr=\frac{BD}{Dc},} with \eqn{BD} is the band depth calculated by
\code{\link[=transformSpeclib]{transformSpeclib()}} and Dc is the maximum band depth called band
centre.  Method \code{"ndbi"} calculates the the normalised band depth index
as \deqn{ndbi= \frac{BD-Dc}{BD+Dc}.} Method \code{"bna"} calculates the band
depth normalised to band area as \deqn{bna=\frac{BD}{Da},} where \eqn{Da} is
the area of the absorption feature (see \code{\link[=feature_properties]{feature_properties()}}).
For further information see Mutanga and Skidmore (2004).
}
\examples{

\% \dontrun{
data(spectral_data)

## Transform speclib
bd <- transformSpeclib(subset(spectral_data, season == "summer"),
                       method = "sh", out = "bd")

## Define features automatically
features <- define.features(bd)

## Isolate the features around 450nm, 700nm, 1200nm and 1500nm and 
## convert to specfeat.
featureSelection <- specfeat(features, c(450,700,1200,1500))

## Plot features
plot(featureSelection,1:4)

## Calculate normalized band depth index for first feature
featureSelection_bdri <- bdri(featureSelection, 1, index = "ndbi")

## Plot result
plot(featureSelection_bdri)
\% }

}
\references{
Mutanga, O. and Skidmore, A. (2004): Hyperspectral band depth
analysis for a better estimation of grass biomass (\var{Cenchrus ciliaris})
measured under controlled laboratory conditions. International Journal of
applied Earth Observation and Geoinformation, 5, 87-96
}
\seealso{
\code{\link[=transformSpeclib]{transformSpeclib()}}, \code{\link[=define.features]{define.features()}},
\code{\link[=specfeat]{specfeat()}}
}
\author{
Lukas Lehnert and Hanna Meyer
}
\keyword{utilities}
