% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{methods}
\name{caret::createDataPartition-methods}
\alias{caret::createDataPartition-methods}
\alias{createDataPartition-methods}
\alias{createDataPartition,ANY-method}
\alias{createDataPartition,.CaretHyperspectral-method}
\title{Methods for Function \code{createDataPartition}}
\description{
Methods for function \code{createDataPartition} in package \pkg{caret}
}
\section{Methods}{
 \describe{

\item{list("signature(y = ".CaretHyperspectral")")}{ Wrapper method for
\code{\link[=createDataPartition]{createDataPartition()}}.  \cr Note that
\code{".CaretHyperspectral"} is a class union containing classes \Sexpr{
paste(hsdar:::.getCaretCompatibleClasses(), collapse = ", ") }. } }
}

\keyword{methods}
