% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{methods}
\name{caret::createFolds-methods}
\alias{caret::createFolds-methods}
\alias{createFolds-methods}
\alias{createMultiFolds-methods}
\alias{createFolds,ANY-method}
\alias{createMultiFolds,ANY-method}
\alias{createFolds,.CaretHyperspectral-method}
\alias{createMultiFolds,.CaretHyperspectral-method}
\title{Methods for Function \code{createFolds} and \code{createMultiFolds}}
\description{
Methods for functions \code{createFolds} and \code{createMultiFolds} in
package \pkg{caret}
}
\section{Methods}{
 \describe{

\item{list("signature(y = ".CaretHyperspectral")")}{ Wrapper methods for
\code{\link[=createFolds]{createFolds()}} and \code{\link[=createMultiFolds]{createMultiFolds()}}.  \cr Note that
\code{".CaretHyperspectral"} is a class union containing classes \Sexpr{
paste(hsdar:::.getCaretCompatibleClasses(), collapse = ", ") }. } }
}

\keyword{methods}
