% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{methods}
\name{caret::createResample-methods}
\alias{caret::createResample-methods}
\alias{createResample-methods}
\alias{createResample,ANY-method}
\alias{createResample,.CaretHyperspectral-method}
\title{Methods for Function \code{createResample}}
\description{
Methods for function \code{createResample} in package \pkg{caret}
}
\section{Methods}{
 \describe{

\item{list("signature(y = ".CaretHyperspectral")")}{ Wrapper method for
\code{\link[=createResample]{createResample()}}.  \cr Note that \code{".CaretHyperspectral"} is
a class union containing classes \Sexpr{
paste(hsdar:::.getCaretCompatibleClasses(), collapse = ", ") }. } }
}

\keyword{methods}
