% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{methods}
\name{caret::featurePlot-methods}
\alias{caret::featurePlot-methods}
\alias{featurePlot-methods}
\alias{featurePlot,ANY-method}
\alias{featurePlot,.CaretHyperspectral-method}
\title{Methods for Function \code{featurePlot}}
\description{
Methods for function \code{featurePlot} in package \pkg{caret}
}
\section{Methods}{
 \describe{

\item{list("signature(x = ".CaretHyperspectral")")}{ Wrapper method for
\code{\link[=featurePlot]{featurePlot()}}.  \cr Note that \code{".CaretHyperspectral"} is a
class union containing classes \Sexpr{
paste(hsdar:::.getCaretCompatibleClasses(), collapse = ", ") }. } }
}

\keyword{methods}
