% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{methods}
\name{caret::preProcess-methods}
\alias{caret::preProcess-methods}
\alias{preProcess-methods}
\alias{preProcess,ANY-method}
\alias{preProcess,.CaretHyperspectral-method}
\alias{show,.preProcessHyperspectral-method}
\alias{preProcess-class}
\title{Methods for Function \code{preProcess}}
\description{
Methods for function \code{preProcess} in package \pkg{caret}. The function
is mainly internally required.
}
\section{Methods}{
 \describe{

\item{list("signature(x = ".CaretHyperspectral")")}{ Wrapper method for
\code{\link[=preProcess]{preProcess()}}.  \cr Note that \code{".CaretHyperspectral"} is a
class union containing classes \Sexpr{
paste(hsdar:::.getCaretCompatibleClasses(), collapse = ", ") }. } }
}

\keyword{methods}
