% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{methods}
\name{caret::rfe}
\alias{caret::rfe}
\alias{rfe-methods}
\alias{get_rfe}
\alias{rfe,Speclib-method}
\alias{rfe,Nri-method}
\alias{rfe,Specfeat-method}
\title{Methods for Function \code{rfe}}
\arguments{
\item{x}{Object of class \code{Speclib}, \code{Nri} or \code{Specfeat}. For
\code{get_rfe}, \code{x} must be the output of \code{rfe} as \code{Speclib}
or \code{Nri}.}

\item{y}{A numeric or factor vector containing the outcome for each sample.
If missing, the response variable set by \code{\link[=setResponse]{setResponse()}} is used.}

\item{cutoff}{The cutoff value of the correlation coefficients between
response variables.}

\item{returnData}{Logical. If TRUE, the updated object of \code{x} is
returned, otherwise only the result of \code{\link[=rfe]{rfe()}} is returned.}

\item{...}{Further aruments passed to \code{\link[=rfe]{rfe()}}.}
}
\value{
If \code{returnData == TRUE}, an object of class \code{Speclib} or
\code{Nri}, otherwise an object of class \code{rfe}. Note that if \code{x}
is an object of class \code{Specfeat}, the function returns an object of
class \code{Speclib} containing the relevant transformed band values.
}
\description{
Methods for function \code{rfe} in package \pkg{caret}.
}
\examples{

\dontrun{
data(spectral_data)

## Set response variable (Chlorophyll content)
spectral_data <- setResponse(spectral_data, "chlorophyll")

## Set additional predictor variables from the attributes
spectral_data <- setPredictor(spectral_data, "season")

## Recursive feature selection
## Note that this may take some time!
rfe_res <- rfe(spectral_data)

get_rfe(rfe_res)

plot(get_rfe(rfe_res))
}

}
\seealso{
\code{\link[=rfe]{rfe()}}
}
\author{
Lukas Lehnert
}
\keyword{methods}
