% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{methods}
\name{caret::setPredictor}
\alias{caret::setPredictor}
\alias{setPredictor-methods}
\alias{setPredictor}
\alias{setPredictor,.CaretHyperspectral,character-method}
\title{Set predictor variable(s)}
\arguments{
\item{x}{Object of one of the following classes: \Sexpr{
paste(hsdar:::.getCaretCompatibleClasses(), collapse = ", ") }.}

\item{predictor}{Character vector. Name of additional predictor variable(s)
(from the attributes).}
}
\value{
The updated object.
}
\description{
Set predictor variable(s) to be used in functions of package \pkg{caret}.
}
\examples{

\dontrun{
data(spectral_data)

## Use subset of data
x <- spectral_data[c(1:31),] 

## Set additional predictor variables from the attributes
x <- setPredictor(x, "season")
}

}
\seealso{
\code{\link[=sbf]{sbf()}}
}
\author{
Lukas Lehnert
}
\keyword{methods}
