% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{methods}
\name{caret::setResponse}
\alias{caret::setResponse}
\alias{setResponse-methods}
\alias{setResponse}
\alias{setResponse,.CaretHyperspectral,character-method}
\title{Set response variable}
\arguments{
\item{x}{Object of one of the following classes: \Sexpr{
paste(hsdar:::.getCaretCompatibleClasses(), collapse = ", ") }.}

\item{response}{Character. Name of response variable (from the attributes).}
}
\value{
The updated object.
}
\description{
Set response variable to be used in functions of package \pkg{caret}.
}
\examples{

\dontrun{
data(spectral_data)

## Use subset of data
x <- spectral_data[c(1:31),] 

## Set response variable (Percentage of green vegetation)
x <- setResponse(x, "PV")
}

}
\author{
Lukas Lehnert
}
\keyword{methods}
