% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{methods}
\name{caret::train-methods}
\alias{caret::train-methods}
\alias{train-methods}
\alias{train,ANY-method}
\alias{train,.CaretHyperspectral-method}
\title{Methods for Function \code{train}}
\description{
Methods for functions \code{train} and \code{train.formula} in package
\pkg{caret}
}
\section{Methods}{
 \describe{

\item{list("signature(x = ".CaretHyperspectral")")}{ Wrapper method for
\code{\link[=train]{train()}}.  \cr Note that \code{".CaretHyperspectral"} is a class
union containing classes \Sexpr{ paste(hsdar:::.getCaretCompatibleClasses(),
collapse = ", ") }. } \item{list("signature(form = "formula", data =
"Speclib")")}{ Wrapper method for \code{\link[=train.formula]{train.formula()}} to be used
with objects of class \code{Speclib}. } }
}

\keyword{methods}
