% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speclib_differenciate.R
\name{derivative.speclib}
\alias{derivative.speclib}
\title{Derivation}
\usage{
derivative.speclib(x, m = 1, method = "sgolay", ...)
}
\arguments{
\item{x}{Object of class \code{Speclib}.}

\item{m}{Return the m-th derivative of the spectra.}

\item{method}{Character string giving the method to be used. Valid options
are \code{"finApprox"} or \code{"sgolay"}.}

\item{...}{Further arguments passed to \code{\link[=sgolayfilt]{sgolayfilt()}}.}
}
\value{
Object of class \code{\linkS4class{Speclib}}.
}
\description{
Calculate derivations of spectra
}
\details{
Two different methods are available: \itemize{ \item Finite approximation
(\code{finApprox}):
\deqn{\frac{dr}{d\lambda}=\frac{r(\lambda_i)-r(\lambda_{i+1})}{\Delta\lambda},}
where \eqn{r_i} is the reflection in band \eqn{i} and \eqn{\Delta\lambda}
the spectral difference between adjacent bands.  \item Savitzky-Golay
derivative computation (\code{sgolay})%:

% \deqn{\frac{dr^q}{d\lambda}=\sum_{i=-m}^{m}P_i^{(q)}r_{j+i},}

}
}
\examples{

data(spectral_data)

## Calculate 1st derivation
d1 <- derivative.speclib(spectral_data)

## Calculate 2nd derivation
d2 <- derivative.speclib(spectral_data, m = 2)

## Calculate 3rd derivation
d3 <- derivative.speclib(spectral_data, m = 3)

par(mfrow=c(2,2))
plot(spectral_data)
plot(d1)
plot(d2)
plot(d3)

}
\references{
Tsai, F. & Philpot, W. (1998): Derivative analysis of
hyperspectral data. Remote Sensing of Environment 66/1. 41-51.
}
\seealso{
\code{\link[=sgolayfilt]{sgolayfilt()}}, \code{\link[=vegindex]{vegindex()}}
}
\author{
Lukas Lehnert
}
\keyword{utilities}
