% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{methods}
\name{distMat3D}
\alias{distMat3D}
\alias{show,DistMat3D-method}
\alias{as.array,DistMat3D-method}
\alias{as.matrix,DistMat3D-method}
\alias{dim,DistMat3D-method}
\alias{ncol,DistMat3D-method}
\alias{nrow,DistMat3D-method}
\alias{distMat3D,matrix-method}
\alias{distMat3D,numeric-method}
\alias{distMat3D,array-method}
\alias{[,DistMat3D,ANY,ANY-method}
\alias{[,DistMat3D,ANY,ANY,ANY-method}
\alias{[<-,DistMat3D,ANY,ANY-method}
\alias{[<-,DistMat3D,ANY,ANY,ANY-method}
\alias{<,DistMat3D,ANY-method}
\alias{<=,DistMat3D,ANY-method}
\alias{>,DistMat3D,ANY-method}
\alias{>=,DistMat3D,ANY-method}
\alias{==,DistMat3D,ANY-method}
\alias{<,DistMat3D,DistMat3D-method}
\alias{<=,DistMat3D,DistMat3D-method}
\alias{>,DistMat3D,DistMat3D-method}
\alias{>=,DistMat3D,DistMat3D-method}
\alias{==,DistMat3D,DistMat3D-method}
\alias{<,ANY,DistMat3D-method}
\alias{<=,ANY,DistMat3D-method}
\alias{>,ANY,DistMat3D-method}
\alias{>=,ANY,DistMat3D-method}
\alias{==,ANY,DistMat3D-method}
\title{Methods to create, manipulate and query objects of class 'DistMat3D'.}
\arguments{
\item{x, object}{Matrix, numeric or array in cases of creation of 'DistMat3D'
objects otherwise object of class 'DistMat3D'.}

\item{ncol}{Number of columns in the new 'DistMat3D' object.}

\item{nlyr}{Number of layer in the new 'DistMat3D' object.}

\item{lower_tri}{Flag if only the lower triangle is used.}

\item{lyr}{Layer in the 'DistMat3D' object to be transformed into matrix.}

\item{value}{Object of class numeric, matrix or array which is used for
replacement of the values in x.}

\item{i, j, n}{Subscripts to access data.}
}
\description{
Methods to create, manipulate and query objects of class 'DistMat3D'. The
following relational operators are defined to compare values between
'DistMat3D'-object(s): \code{<, <=, ==, >, >=}
}
\examples{

data(spectral_data)

## Mask channel crossing part (arround 1050 nm) and strong 
## water absorption part (above 1350 nm)
mask(spectral_data) <- c(1045, 1055, 1350, 1706)

## Calculate SAM distances (object of class 'dist')
sam_dist <- dist.speclib(subset(spectral_data, season == "summer"))

## Convert to class 'distMat3D'
sam_dist <- distMat3D(as.matrix(sam_dist))

sam_dist

}
\seealso{
\code{\linkS4class{DistMat3D}},
\code{\link[=apply.DistMat3D:apply]{=apply.DistMat3D::apply()}}, \code{\linkS4class{Nri}}
}
\author{
Lukas Lehnert
}
\keyword{classes}
