% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_resampling_response.R
\name{get.response}
\alias{get.response}
\alias{get_RapidEye_response}
\alias{get_response_function}
\title{Satellite response functions}
\usage{
get.response(sensor, range = NULL, response_function = TRUE,
  continuousdata = "auto")
}
\arguments{
\item{sensor}{Name or integer value of satellite sensor. Matched with
\code{\link[=get.sensor.name]{get.sensor.name()}}.}

\item{range}{Vector of length = 2 containing maximum and minimum wavelength
to be considered.}

\item{response_function}{If \code{TRUE}, spectral response function else
wise Gaussian response function will be returned.}

\item{continuousdata}{Definition if returned \code{\linkS4class{Speclib}} is
containing continuous data or not.}
}
\value{
Object of class \code{\linkS4class{Speclib}} containing spectral
response values instead of reflectance value. Spectral response values may
be accessed with \code{\link[=spectra.Speclib:spectra]{=spectra.Speclib::spectra()}}.
}
\description{
Handling satellite sensor response functions
}
\examples{

## See example in get.sensor.characteristics

}
\seealso{
\code{\link[=get.sensor.name]{get.sensor.name()}},
\code{\link[=get.sensor.characteristics]{get.sensor.characteristics()}},
\code{\link[=get.gaussian.response]{get.gaussian.response()}}
}
\author{
Lukas Lehnert
}
\keyword{utilities}
