% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_resampling_sensorCharacteristics.R
\name{get.sensor.characteristics}
\alias{get.sensor.characteristics}
\alias{list.available.sensors}
\title{Sensor characteristics}
\usage{
get.sensor.characteristics(sensor, response_function = FALSE)
}
\arguments{
\item{sensor}{Character or integer. Name or numerical abbreviation of
sensor. See 'sensor="help"' or 'sensor=0' for an overview of available
sensors.}

\item{response_function}{If TRUE, the spectral response function is returned}
}
\description{
Get channel wavelength of satellite sensor
}
\examples{

## Return implemented sensors
get.sensor.characteristics(0)

## RapidEye
data_wv <- get.sensor.characteristics("RapidEye", TRUE)

## Plot response functions
plot(c(0,1)~c(330,1200), type = "n", xlab = "Wavelength [nm]", 
     ylab = "Spectral response")
xwl_response <- seq.int(attr(data_wv$response, "minwl"),
                        attr(data_wv$response, "maxwl"),
                        attr(data_wv$response, "stepsize"))
for (i in 1:nrow(data_wv$characteristics))
  lines(xwl_response, data_wv$response[,i], col = i)

}
\seealso{
\code{\link[=spectralResampling]{spectralResampling()}}
}
\author{
Lukas Lehnert
}
\keyword{utilities}
