% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_resampling_sensorCharacteristics.R
\name{get.sensor.name}
\alias{get.sensor.name}
\title{Satellite sensor name}
\usage{
get.sensor.name(sensor)
}
\arguments{
\item{sensor}{Integer value to match against predefined satellite sensors.}
}
\value{
Name of satellite sensor as character string.
}
\description{
Get satellite sensor name by integer value
}
\details{
See \code{\link[=get.sensor.characteristics]{get.sensor.characteristics()}} to get overview on available
satellite sensors.
}
\examples{

get.sensor.name(1)

}
\seealso{
\code{\link[=get.sensor.characteristics]{get.sensor.characteristics()}}
}
\author{
Lukas Lehnert
}
\keyword{utilities}
