% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{package}
\name{hsdar-package}
\alias{hsdar-package}
\alias{hsdar}
\title{Manage, analyse and simulate hyperspectral data in R}
\description{
The \pkg{hsdar} package contains classes and functions to manage, analyse
and simulate hyperspectral data. These might be either spectrometer
measurements or hyperspectral images through the interface of \pkg{raster}.
}
\section{Acknowledgements}{
 Development initially funded by German Federal
Ministry of Education and Research (03G0808C) in the scope of the project
PaDeMoS as precondition to develop a space-based Pasture Degradation
Monitoring System for the Tibetan Plateau.
}

\author{
Lukas Lehnert, Hanna Meyer, Joerg Bendix
}
\keyword{package}
