% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdardocs.R
\name{hsdardocs}
\alias{hsdardocs}
\title{Load additional documents}
\usage{
hsdardocs(doc)
}
\arguments{
\item{doc}{Name of document to load. Currently, only \code{"References.pdf"}
and \code{"Copyright"} are available}
}
\description{
Access help documents and references for different methods.
}
\examples{

\dontrun{
## Open references of hyperspectral vegetation indices (PDF-file)
hsdardocs("References.pdf")

## See copyrights of routines and data used in hsdar-package (ascii-file)
hsdardocs("Copyright")
}

}
\author{
Lukas Lehnert
}
\keyword{documentation}
\keyword{utilities}
