% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speclib_id.R
\name{idSpeclib}
\alias{idSpeclib}
\alias{idSpeclib<-}
\title{Handling IDs of spectra}
\usage{
idSpeclib(x)
}
\arguments{
\item{x}{Object of class \code{Speclib}.}

\item{value}{Character vector of the same length as \code{nspectra(x)}, or
NULL.}
}
\value{
For \code{idSpeclib<-}, the updated object. Otherwise a vector
giving the ID of each spectrum in Speclib is returned.
}
\description{
Returning and setting ID of spectra in Speclib
}
\examples{

data(spectral_data)

idSpeclib(spectral_data)

}
\seealso{
\code{\linkS4class{Speclib}}
}
\author{
Lukas Lehnert
}
\keyword{utilities}
