% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuum_features.R
\docType{methods}
\name{plot,Specfeat,ANY-method}
\alias{plot,Specfeat,ANY-method}
\alias{attribute}
\alias{attribute.speclib}
\alias{attribute<-}
\alias{attribute,Speclib-method}
\alias{attribute<-,Speclib,data.frame-method}
\alias{attribute<-,Speclib,matrix-method}
\alias{attribute,Nri-method}
\alias{attribute<-,Nri,data.frame-method}
\alias{attribute<-,Nri,matrix-method}
\title{Handling attributes of spectra}
\usage{
\S4method{plot}{Specfeat,ANY}(x, fnumber = 1, stylebysubset = NULL,
  changecol = TRUE, changetype = FALSE, autolegend = TRUE, new = TRUE,
  ...)
}
\arguments{
\item{object}{Object of class \code{Speclib} or \code{Nri}.}

\item{value}{Data frame with \code{nrow(value) == nspectra(object)}, or
NULL.}
}
\value{
For \code{attribute<-}, the updated object. \code{attribute} returns
a data frame with attribute data.% The return value of \code{setattributes}
is an object of class speclib.
}
\description{
Returning and setting attributes of spectra in Speclib or Nri.%
\code{setattributes} is a more comfortable and save function to set
attributes of speclib using string matching.
}
\examples{

data(spectral_data)

## Returning attributes
attribute(spectral_data)

}
\seealso{
\code{\linkS4class{Speclib}}, \code{\linkS4class{Nri}}
}
\author{
Lukas Lehnert
}
\keyword{utilities}
