% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastermeta.R
\name{rastermeta}
\alias{rastermeta}
\title{Create list containing rastermeta-information}
\usage{
rastermeta(x, dim, ext, crs)
}
\arguments{
\item{x}{Optional. Object of one of the following classes: "Raster",
"RasterBrick", "RasterStack", "HyperSpecRaster".}

\item{dim}{Optional. Vector with length == 2. The first and second elements
give the number of rows and columns, respectively.}

\item{ext}{Optional. Object of class \code{extent}.}

\item{crs}{Optional. Object of class \code{CRS}.}
}
\value{
List with following elements (in exactly this order!): \itemize{
\item dim: Vector with length == 2. The first and second elements give the
number of rows and columns, respectively.  \item ext: Object of class
\code{extent}.  \item crs: Object of class \code{CRS}.  }
}
\description{
Create valid objects for slot \code{rastermeta} in
\code{\linkS4class{Speclib}}.
}
\seealso{
\code{\linkS4class{Speclib}}, \code{\linkS4class{HyperSpecRaster}}
}
\author{
Lukas Lehnert
}
