\name{read_header}
\alias{read_header}


\title{Get reflectance values}
\description{
Read ENVI header file
}

\usage{
read_header(file, ...)

}
\arguments{
  \item{file}{
    Path of file to be read.
  }
  \item{...}{
    Arguments to be passed to specific functions. Currently ignored.
  }
}
\value{
  A named list containing the information in the header file
}
\author{ Lukas Lehnert }

