% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{return_index}
\alias{return_index}
\title{returns vegetation indeces}
\usage{
return_index(x)
}
\description{
returns vegetation indeces
useless since it does not do anything because 'x' is just returned ???

Used in \code{vegindex}.
}
\keyword{internal}
