% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegindex.R
\name{vegindex}
\alias{vegindex}
\title{vegindex}
\usage{
vegindex(x, index, returnHCR = "auto", L = 0.5, weighted = TRUE,
  bnames = NULL, filename = NULL, method = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{Speclib} or \code{HyperSpecRaster}.}

\item{index}{Character string. Name or definition of index or vector with
names/definitions of indices to calculate. See Details section for further
information.}

\item{returnHCR}{If TRUE, the result will be of class HyperSpecRaster,
otherwise it is a data frame. If "auto", the class is automatically
determined by passed Speclib.}

\item{L}{Factor for SAVI index. Unused for other indices.}

\item{weighted}{Logical indicating if reflectance values should be
interpolated to fit wavelength position. If \code{FALSE} the reflectance
values of nearest neighbour to passed position are returned. See
\code{\link[=get_reflectance.speclib:get_reflectance]{=get_reflectance.speclib::get_reflectance()}} for further
explanation.}

\item{bnames}{(optional) Character vector of band names. Only used if
an object of class \code{HyperSpecRaster} is provided.}

\item{filename}{(optional) Filename of the raster file written to disk. Only used if
an object of class \code{HyperSpecRaster} is provided.}

\item{method}{Character string giving the method to be used in \code{derivate_speclib}.
Valid options are "finApprox" or "sgolay". See details for more.}

\item{...}{Further arguments passed to derivative functions. Only used for
indices requiring derivations.}
}
\value{
A vector containing indices values. If index is a vector with
\code{length > 1}, a data frame with \code{ncol = length(index)} and
\code{nrow = number of spectra in x} is returned.

If function is called without any arguments, return value will be a vector
containing all available indices in alphabetical order.
}
\description{
Function calculates a variety of hyperspectral vegetation indices
}
\examples{

## Example calculating NDVI
data(spectral_data)
ndvi <- vegindex(spectral_data, "NDVI")


## Example calculating all available indices
## Get available indices
avl <- vegindex()
vi <- vegindex(spectral_data, avl)

## Self-defined indices
## NDVI
vi <- vegindex(spectral_data, "(R800-R680)/(R800+R680)")
## NDNI
vi <- vegindex(spectral_data, 
               "(log(1/R1510) - log(1/R1680))/(log(1/R1510) + log(1/R1680))")
## D1
vi <- vegindex(spectral_data, "D730/D706")

## HyperSpecRaster example

data("hyperspecs") 
out <- vegindex(hyperspecs, "NDVI", method = "finApprox", nl = 1)

}
\references{
See \code{hsdardocs("References.pdf")}
}
\seealso{
\code{\link[=soilindex]{soilindex()}}, \code{\link[=derivative.speclib]{derivative.speclib()}},
\code{\link[=rededge]{rededge()}},
\code{\link[=get_reflectance.speclib:get_reflectance]{=get_reflectance.speclib::get_reflectance()}}
}
\author{
Hanna Meyer and Lukas Lehnert
}
\keyword{multivariate}
