% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{vegindex_available}
\alias{vegindex_available}
\title{Returns available vegetation indices}
\usage{
vegindex_available()
}
\description{
Returns available vegetation indices

Used in \code{vegindex}.
}
\keyword{internal}
