% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuum_clman.R
\name{addcp}
\alias{addcp}
\title{Add fix point}
\usage{
addcp(x, ispec, cpadd)
}
\arguments{
\item{x}{Object of class Clman.}

\item{ispec}{ID or index of spectrum to be modified.}

\item{cpadd}{Single value or vector of wavelength containing new fix points.}
}
\value{
Object of class \code{Clman} containing the updated version of x.
}
\description{
Add fix point to continuum line.
}
\examples{

## Model spectra using PROSAIL
parameter <- data.frame(N = rep.int(c(1, 1.5),2), LAI = c(1,1,3,3))
spec <- PROSAIL(parameterList=parameter)

## Transform spectra
spec_clman <- transformSpeclib(spec, method = "sh", out = "raw")

## Plot original line
par(mfrow = c(1,2))
plot(spec_clman, ispec = 1, subset = c(2480, 2500))

## Add fix point at 4595 nm to continuum line of first spectrum
spec_clman <- addcp(spec_clman, 1, 2495)

## Plot new line
plot(spec_clman, ispec = 1, subset = c(2480, 2500))

## Check new hull
hull <- checkhull(spec_clman, 1)
hull$error

}
\seealso{
\code{\link[=transformSpeclib]{transformSpeclib()}}, \code{\link[=deletecp]{deletecp()}},
\code{\link[=getcp]{getcp()}}, \code{\link[=checkhull]{checkhull()}}, \code{\link[=makehull]{makehull()}},
\code{\link[=updatecl]{updatecl()}},

\code{\link[=idSpeclib]{idSpeclib()}}
}
\author{
Lukas Lehnert and Hanna Meyer
}
\keyword{utilities}
