\name{cancer_spectra}
\alias{cancer_spectra}
\docType{data}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Hyperspectral samples
}
\description{
Hyperspectral samples from the human larynx
}
\usage{
data(cancer_spectra)
}
\details{
  BIANCA
}
\format{An object of class \code{Speclib}}
\author{Bianca Regeling, Lukas Lehnert}

\keyword{datasets}

