% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{data}
\name{cancer_spectra}
\alias{cancer_spectra}
\title{Hyperspectral samples}
\format{An object of class \code{Speclib}}
\description{
Hyperspectral samples from the human larynx
}
\details{
BIANCA
}
\author{
Bianca Regeling, Lukas Lehnert
}
\keyword{datasets}
