% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuum_transform.R
\name{checkhull}
\alias{checkhull}
\title{Check continuum line}
\usage{
checkhull(x, ispec)
}
\arguments{
\item{x}{Object of class \code{clman}.}

\item{ispec}{ID or index of spectrum to be checked.}
}
\value{
Object of class \code{list}.
}
\description{
Check if continuum line is intersecting the reflectance curve.
}
\examples{

## Model spectra using PROSAIL
parameter <- data.frame(N = rep.int(c(1, 1.5),2), LAI = c(1,1,3,3))
spec <- PROSAIL(parameterList=parameter)

## Transform spectra
spec_clman <- transformSpeclib(spec, method = "sh", out = "raw")

## Plot original line
par(mfrow = c(1,2))
plot(spec_clman, ispec = 1, subset = c(2480, 2500))

## Add fix point at 4595 nm to continuum line of first spectrum
spec_clman <- addcp(spec_clman, 1, 2495)

## Plot new line
plot(spec_clman, ispec = 1, subset = c(2480, 2500))

## Check new hull
hull <- checkhull(spec_clman, 1)
hull$error

## Add fix point at 4596 nm to continuum line of first spectrum
spec_clman <- addcp(spec_clman, 1, 2496)

## Check new hull
hull <- checkhull(spec_clman, 1)
hull$error

}
\seealso{
\code{\link[=transformSpeclib]{transformSpeclib()}}, \code{\link[=addcp]{addcp()}},
\code{\link[=deletecp]{deletecp()}}, \code{\link[=makehull]{makehull()}}, \code{\link[=updatecl]{updatecl()}}
}
\author{
Lukas Lehnert and Hanna Meyer
}
\keyword{utilities}
