% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{class}
\name{Clman}
\alias{Clman}
\alias{plot,Clman-method}
\title{Clman class}
\arguments{
\item{wavelength}{Vector with corresponding wavelength for each band.}

\item{cp}{Data frame or matrix containing fix points. Fix points have
numbers greater than 0, all other bands are 0.}

\item{hull}{Data frame or matrix containing linear hull.}

\item{spectra}{Data frame, matrix of raster object of class
'SpatialGridDataFrame' with spectral data.}

\item{outdatedhull}{Data frame or matrix containing hull of step before for
undo porposes.}

\item{mask}{Data frame with masked parts in the spectra. See
\code{\link[=mask]{mask()}}.}

\item{x}{Object of class \code{clman}.}

\item{ispec}{Name or index of spectrum to be plotted.}

\item{subset}{Lower and upper spectral limits used for plot.}

\item{numeratepoints}{Flag if points should be numerated in plot.}

\item{hull.style}{List of arguments passed to \code{\link[=lines]{lines()}} to
construct the continuum line.}

\item{points.style}{List of arguments passed to \code{\link[=points]{points()}} to
construct the continuum points. May be \code{NULL} to suppress plotting of
fix points.}

\item{...}{Further arguments passed to plot.default.}
}
\value{
Object of class \code{\linkS4class{Clman}}.
}
\description{
Class to store and handle manual continuum lines
}
\examples{

## Model spectra using PROSAIL
parameter <- data.frame(N = rep.int(c(1, 1.5),2), LAI = c(1,1,3,3))
spec <- PROSAIL(parameterList=parameter)

## Transform spectra
spec_clman <- transformSpeclib(spec, method = "sh", out = "raw")

## Plot clman
plot(spec_clman, ispec = 1, subset = c(400, 1000))

}
\seealso{
\code{\link[=transformSpeclib]{transformSpeclib()}}, \code{\link[=plot.Speclib:plot]{=plot.Speclib::plot()}}
}
\author{
Lukas Lehnert and Hanna Meyer
}
\keyword{aplot}
\keyword{classes}
