% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{methods}
\name{clman}
\alias{clman}
\alias{initialize,Clman-method}
\alias{spectra,Clman-method}
\alias{spectra<-,Clman,data.frame-method}
\alias{spectra<-,Clman,matrix-method}
\alias{spectra<-,Clman,numeric-method}
\title{Methods to create, manipulate and query objects of class 'Clman'.}
\arguments{
\item{.Object, object}{Matrix, numeric or array in cases of creation of
'Clman' objects otherwise object of class 'Clman'.}

\item{value}{Object of class numeric, matrix or array which is used for
replacement of the values in x.}

\item{...}{Arguments passed to \code{\link[=createspeclib]{createspeclib()}}.}
}
\value{
For \code{spectra<-}, the updated object. Otherwise a matrix
returning the spectra in the Clman object.
}
\description{
Methods to create, manipulate and query objects of class 'Clman'.
}
\note{
The functions to create objects of class \code{Clman} are mainly
internally needed by \code{\link[=transformSpeclib]{transformSpeclib()}}.
}
\seealso{
\code{\link[=dist.speclib]{dist.speclib()}}, \code{\link[=Clman]{Clman()}},
\code{\link[=transformSpeclib]{transformSpeclib()}}, \code{\link[=plot.Speclib:plot]{=plot.Speclib::plot()}}
}
\author{
Lukas Lehnert
}
