% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{class}
\name{DistMat3D-class}
\alias{DistMat3D-class}
\title{\itemize{
\item DistMat3D class
}}
\description{
Class to store effectively (large) distance matrices (up to 3D).
}
\details{
Object with 3 slots: \itemize{ \item values: Numerical vector containing
distance values \item ncol: Number of columns in the 3D-matrix. Number of
columns equals always the number of rows \item nlyr: Number of layers in the
3D-matrix }
}
\note{
See figure in \code{\link[=hsdar-package]{hsdar-package()}} for an overview of classes
in hsdar.
}
\seealso{
\code{\link[=distMat3D]{distMat3D()}}
}
\author{
Lukas Lehnert
}
\keyword{classes}
