% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicore_funtions.R
\name{hsdar_parallel}
\alias{hsdar_parallel}
\title{hsdar_parallel}
\usage{
hsdar_parallel()
}
\value{
Vector containing supported function names
}
\description{
Get all functions which support parallel execution
}
\details{
Parallel execution is performed via the \pkg{foreach}-package. Care is taken
that a function will never run in parallel if the calling function is
already using multicore processing.
}
\examples{

\dontrun{
data(spectral_data)
## Load library
library(doMC)
## Register number of workers
registerDoMC(3)

supported_functions <- hsdar_parallel()
supported_functions

## Transform speclib using 3 cores
bd <- transformSpeclib(spectral_data)
}

}
\author{
Lukas Lehnert
}
\keyword{utilities}
