% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speclib_smooth.R
\name{meanfilter}
\alias{meanfilter}
\title{Apply mean filter}
\usage{
meanfilter(spectra, p = 5)
}
\arguments{
\item{spectra}{Data frame containing spectra}

\item{p}{Filter size.}
}
\value{
Filtered data frame of same dimension as input data frame
}
\description{
Apply mean filter to data frame with spectra as rows and bands as columns.
Filter size is passed as number of bands averaged at both sites of the
respective band value.
}
\examples{

data(spectral_data)

spectra_filtered <- meanfilter(spectra(spectral_data), p = 10)
spectra(spectral_data) <- spectra_filtered 

  

}
\seealso{
\code{\link[=smoothSpeclib]{smoothSpeclib()}}
}
\author{
Lukas Lehnert
}
\keyword{smooth}
