% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbi_raster.R
\name{nbi_raster}
\alias{nbi_raster}
\title{Calculates all possible combinations of narrow band indices}
\usage{
nbi_raster(x, weighted = TRUE, bnames_prefix = "", filename = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{HyperSpecRaster}.}

\item{weighted}{Logical indicating if reflectance values should be
interpolated to fit wavelength position. If \code{FALSE} the reflectance
values of nearest neighbour to passed position are returned. See
\code{\link[=get_reflectance]{get_reflectance()}} for further
explanation.}

\item{bnames_prefix}{(optional) Prefix of band name}

\item{filename}{(optional) Filename of the raster file written to disk.}
}
\value{
Object of class \code{RasterBrick}
}
\description{
Calculates all possible combinations of narrow band indices
}
\examples{

data(hyperspecs)
hyperspecs@file@name = system.file("data/laukiz1.tif", package = "hsdar")
nbi <- nbi_raster(hyperspecs)

}
\seealso{
\code{\link[=vegindex]{vegindex()}}, \code{\link[=nri]{nri()}}, \code{\link[=get_reflectance]{get_reflectance()}}
}
\author{
Patrick Schratz
}
