\name{Specfeat-class}
\docType{class}
\alias{Specfeat-class}

\title{
* Specfeat class
}
\description{
Class to handle spectral feature data.
}
\details{
Class extends \code{Speclib}-class and adds two additional slots:
\itemize{
  \item{features: }{List containing the spectra according to the features.}
  \item{featureLimits: }{List containing limits of features defined by \code{\link{define.features}}.}
}
}
\note{
See figure in \code{\link{hsdar-package}} for an overview of classes in hsdar.
}

\author{
Lukas Lehnert
}


\seealso{
\code{\linkS4class{Speclib}}
}
\keyword{classes}
