% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{class}
\name{Specfeat-class}
\alias{Specfeat-class}
\title{\itemize{
\item Specfeat class
}}
\description{
Class to handle spectral feature data.
}
\details{
Class extends \code{Speclib}-class and adds two additional slots: \itemize{
\item features: List containing the spectra according to the features.
\item featureLimits: List containing limits of features defined by
\code{\link[=define.features]{define.features()}}. }
}
\note{
See figure in \code{\link[=hsdar-package]{hsdar-package()}} for an overview of classes
in hsdar.
}
\seealso{
\code{\linkS4class{Speclib}}
}
\author{
Lukas Lehnert
}
\keyword{classes}
