\name{spectral_data}
\alias{spectral_data}
\docType{data}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Hyperspectral samples
}
\description{
Hyperspectral samples from a FACE experiment in Germany
}
\usage{
data(spectral_data)
}
\details{
  Data has been sampled during vegetation period 2014 in spring and summer. Measurements were taken with a HandySpec Field portable spectrometer (tec5 AG Oberursel, Germany). This device has two channels measuring incoming and reflected radiation simultaneously between 305 and 1705 nm in 1 nm steps.
}
\format{An object of class \code{Speclib}}
\author{Wolfgang A. Obermeier, Lukas Lehnert, Hanna Meyer}

\keyword{datasets}

