% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsdar-package.R
\docType{data}
\name{spectral_data}
\alias{spectral_data}
\title{Hyperspectral samples}
\format{An object of class \code{Speclib}}
\description{
Hyperspectral samples from a FACE experiment in Germany
}
\details{
Data has been sampled during vegetation period 2014 in spring and summer.
Measurements were taken with a HandySpec Field portable spectrometer (tec5
AG Oberursel, Germany). This device has two channels measuring incoming and
reflected radiation simultaneously between 305 and 1705 nm in 1 nm steps.
}
\author{
Wolfgang A. Obermeier, Lukas Lehnert, Hanna Meyer
}
\keyword{datasets}
